requirements.txt
fastapi
uvicorn
streamlit
psycopg2-binary
SQLAlchemy
pandas
plotly
numpy
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
pdfplumber
pdfminer.six
pypdf
openai
langchain
langchain-core
langchain-openai
langchain-community
python-dotenv
pydantic
sentence-transformers
chromadb
scikit-learn
passlib
bcrypt
PyJWT
